requirements.txt
gpsbabel
pyserial
python2-biggles
pyudev
folium